+++ Common/src/main/java/ca/lukegrahamlandry/lib/base/json/JsonHelper.java
@@ -13,6 +13,7 @@ import ca.lukegrahamlandry.lib.base.GenericHolder;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.TypeAdapterFactory;
+import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
@@ -26,6 +27,7 @@ public class JsonHelper {
             .registerTypeAdapter(GenericHolder.class, new GenericHolder.TypeAdapter())
             .registerTypeAdapter(CompoundTag.class, new NbtTypeAdapter())
             .registerTypeAdapter(ItemStack.class, new ItemStackTypeAdapter())
+            .registerTypeAdapter(BlockPos.class, new BlockPosTypeAdapter())
             .registerTypeAdapterFactory(new RegistryObjectTypeAdapterFactory());
 
     private static Gson GSON = GSON_BUILDER.create();