+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 
 ## 1.1.0
 
-- ConfigWrapper can load a List or Map as the top level object
-- KeybindWrapper
+- [X] ConfigWrapper can load a List or Map as the top level object
+- [X] KeybindWrapper
+- [X] PlatformHelper: isModLoaded
+- [ ] data packs
+- [ ] config reloading
+- [ ] config array / map comments on fields brought to top if list empty
+- [ ] nbt data for tile, entity, item
+- [ ] quilt example mod
 
 ## 1.0.0